Bug 757411 - Add Installer/Uninstaller improvements from Mozmill automation scripts to mozInstall. r=jhammel

+++ mozinstall/mozinstall/mozinstall.py
@@ -4,6 +4,10 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this file,
 # You can obtain one at http://mozilla.org/MPL/2.0/.
 
+"""Module to handle the installation and uninstallation of Gecko based
+applications across platforms.
+
+"""
 import mozinfo
 from optparse import OptionParser
 import os
@@ -11,59 +15,109 @@ import shutil
 import subprocess
 import sys
 import tarfile
+import time
 import zipfile
 
 if mozinfo.isMac:
     from plistlib import readPlist
 
 
-_default_apps = ['firefox',
-                 'thunderbird',
-                 'fennec']
+DEFAULT_APPS = ['firefox',
+                'thunderbird',
+                'fennec']
+
+TIMEOUT_UNINSTALL = 60
 
 
 class InstallError(Exception):
-    """
-    Thrown when the installation fails. Includes traceback
-    if available.
-    """
+    """Thrown when installation fails. Includes traceback if available."""
 
 
 class InvalidBinary(Exception):
-    """
-    Thrown when the binary cannot be found after the installation.
-    """
+    """Thrown when the binary cannot be found after the installation."""
 
 
 class InvalidSource(Exception):
-    """
-    Thrown when the specified source is not a recognized
-    file type (zip, exe, tar.gz, tar.bz2 or dmg)
-    """
+    """Thrown when the specified source is not a recognized file type.
 
+    Supported types:
+    Linux:   tar.gz, tar.bz2
+    Mac:     dmg
+    Windows: zip, exe
 
-def install(src, dest=None, apps=_default_apps):
-    """
-    Installs a zip, exe, tar.gz, tar.bz2 or dmg file
-    src - the path to the install file
-    dest - the path to install to [default is os.path.dirname(src)]
-
-    returns - the full path to the binary in the installed folder
-              or None if the binary cannot be found
     """
 
+
+class UninstallError(Exception):
+    """Thrown when uninstallation fails. Includes traceback if available."""
+
+
+def get_binary(path, apps=DEFAULT_APPS):
+    """Find the binary in the specified path, and return its path. If binary is
+    not found throw an InvalidBinary exception.
+
+    Arguments:
+    path -- the path within to search for the binary
+
+    Keyword arguments:
+    apps -- list of binaries without file extension to look for
+
+    """
+    binary = None
+
+    # On OS X we can get the real binary from the app bundle
+    if mozinfo.isMac:
+        plist = '%s/Contents/Info.plist' % path
+        assert os.path.isfile(plist), '"%s" has not been found.' % plist
+
+        binary = os.path.join(path, 'Contents/MacOS/',
+                              readPlist(plist)['CFBundleExecutable'])
+
+    else:
+        if mozinfo.isWin:
+            apps = [app + '.exe' for app in apps]
+
+        for root, dirs, files in os.walk(path):
+            for filename in files:
+                # os.access evaluates to False for some reason, so not using it
+                if filename in apps:
+                    binary = os.path.realpath(os.path.join(root, filename))
+                    break
+
+    if not binary:
+        # The expected binary has not been found. Make sure we clean the
+        # install folder to remove any traces
+        shutil.rmtree(path)
+
+        raise InvalidBinary('"%s" does not contain a valid binary.' % path)
+
+    return binary
+
+
+def install(src, dest=None, apps=DEFAULT_APPS):
+    """Install a zip, exe, tar.gz, tar.bz2 or dmg file, and return the path of
+    the binary. If binary is not found throw an InstallError exception.
+
+    Arguments:
+    src  -- the path to the install file
+
+    Keyword arguments:
+    dest -- the path to install to (default: sub folder in current working dir)
+    apps -- list of binaries without file extension to look for
+
+    """
     src = os.path.realpath(src)
     if not is_installer(src):
         raise InvalidSource(src + ' is not a recognized file type ' +
                                   '(zip, exe, tar.gz, tar.bz2 or dmg)')
 
     if not dest:
-        dest = os.path.dirname(src)
+        dest = os.getcwd()
 
         # On Windows the installer doesn't create a sub folder in the
-        # destination folder and would clutter the source folder
+        # destination folder and would clutter the current working dir
         if mozinfo.isWin and src.lower().endswith('.exe'):
-            filename = os.path.splitext(os.path.basename(src))[0]
+            filename = os.path.basename(src).split('.')[0]
             dest = os.path.join(dest, filename)
 
     trbk = None
@@ -78,8 +132,8 @@ def install(src, dest=None, apps=_default_apps):
 
     except Exception, e:
         cls, exc, trbk = sys.exc_info()
-        install_error = InstallError('Failed to install %s' % src)
-        raise install_error.__class__, install_error, trbk
+        error = InstallError('Failed to install "%s"' % src)
+        raise InstallError, error, trbk
 
     finally:
         # trbk won't get GC'ed due to circular reference
@@ -91,13 +145,17 @@ def install(src, dest=None, apps=_default_apps):
 
 
 def is_installer(src):
-    """
-    Tests if the given file is a valid installer package (zip, exe, tar.gz,
-    tar.bz2 or dmg)
+    """Tests if the given file is a valid installer package.
 
-    src - the path to the install file
-    """
+    Supported types:
+    Linux:   tar.gz, tar.bz2
+    Mac:     dmg
+    Windows: zip, exe
 
+    Arguments:
+    src -- the path to the install file
+
+    """
     src = os.path.realpath(src)
     assert os.path.isfile(src), 'Installer has to be a file'
 
@@ -109,21 +167,73 @@ def is_installer(src):
         return src.lower().endswith('.exe') or zipfile.is_zipfile(src)
 
 
-def _extract(src, dest=None, delete=False):
+def uninstall(binary):
+    """Uninstalls the specified binary. If it has been installed via an
+    installer on Windows it will make use of the uninstaller first.
+
+    Arguments:
+    binary -- the path to the binary
+
     """
-    Takes in a tar or zip file and extracts it to dest
+    binary = os.path.realpath(binary)
+    assert os.path.isfile(binary), 'binary "%s" has to be a file.' % binary
 
-    If dest is not specified, extracts to os.path.dirname(src)
-    If delete is set to True, deletes the bundle at path
+    # We know that the binary is a file. So we can safely remove the parent
+    # folder. On OS X we have to get the .app bundle.
+    folder = os.path.dirname(binary)
+    if mozinfo.isMac:
+        folder = os.path.dirname(os.path.dirname(folder))
 
-    Returns the list of top level files that were extracted
+    # On Windows we have to use the uninstaller. If it's not available fallback
+    # to the directory removal code
+    if mozinfo.isWin:
+        uninstall_folder = '%s\uninstall' % folder
+        log_file = '%s\uninstall.log' % uninstall_folder
+
+        if os.path.isfile(log_file):
+            trbk = None
+            try:
+                cmdArgs = ['%s\uninstall\helper.exe' % folder, '/S']
+                result = subprocess.call(cmdArgs)
+                if not result is 0:
+                    raise Exception('Execution of uninstaller failed.')
+
+                # The uninstaller spawns another process so the subprocess call
+                # returns immediately. We have to wait until the uninstall
+                # folder has been removed or until we run into a timeout.
+                end_time = time.time() + TIMEOUT_UNINSTALL
+                while os.path.exists(uninstall_folder):
+                    time.sleep(1)
+
+                    if time.time() > end_time:
+                        raise Exception('Failure removing uninstall folder.')
+
+            except Exception, e:
+                cls, exc, trbk = sys.exc_info()
+                error = UninstallError('Failed to uninstall %s' % binary)
+                raise UninstallError, error, trbk
+
+            finally:
+                # trbk won't get GC'ed due to circular reference
+                # http://docs.python.org/library/sys.html#sys.exc_info
+                del trbk
+
+    # Ensure that we remove any trace of the installation. Even the uninstaller
+    # on Windows leaves files behind we have to explicitely remove.
+    shutil.rmtree(folder)
+
+
+def _extract(src, dest):
+    """Extract a tar or zip file into the destination folder and return the
+    application folder.
+
+    Arguments:
+    src -- archive which has to be extracted
+    dest -- the path to extract to
+
     """
 
-    assert not os.path.isfile(dest), "dest cannot be a file"
-
-    if dest is None:
-        dest = os.path.dirname(src)
-    elif not os.path.isdir(dest):
+    if not os.path.exists(dest):
         os.makedirs(dest)
 
     if zipfile.is_zipfile(src):
@@ -161,9 +271,6 @@ def _extract(src, dest=None, delete=False):
 
     bundle.close()
 
-    if delete:
-        os.remove(src)
-
     # namelist returns paths with forward slashes even in windows
     top_level_files = [os.path.join(dest, name) for name in namelist
                              if len(name.rstrip('/').split('/')) == 1]
@@ -177,14 +284,15 @@ def _extract(src, dest=None, delete=False):
     return top_level_files
 
 
-def _install_dmg(src, dest):
-    """
-    Takes in a dmg file and extracts it to destination folder
-    If dest is not specified, extracts to os.path.dirname(src)
+def _install_dmg(src, dest=None):
+    """Extract a dmg file into the destination folder and return the
+    application folder.
 
-    Returns the list of top level files that were extracted
-    """
+    Arguments:
+    src -- DMG image which has to be extracted
+    dest -- the path to extract to
 
+    """
     try:
         proc = subprocess.Popen('hdiutil attach %s' % src,
                                 shell=True,
@@ -203,10 +311,11 @@ def _install_dmg(src, dest):
         mounted_path = os.path.join(appDir, appName)
 
         dest = os.path.join(dest, appName)
-        assert not os.path.isfile(dest)
 
         # copytree() would fail if dest already exists.
-        shutil.rmtree(dest, ignore_errors=True)
+        if os.path.exists(dest):
+            raise InstallError('App bundle "%s" already exists.' % dest)
+
         shutil.copytree(mounted_path, dest, False)
 
     finally:
@@ -217,8 +326,18 @@ def _install_dmg(src, dest):
 
 
 def _install_exe(src, dest):
-    # possibly gets around UAC in vista (still need to run as administrator)
+    """Run the MSI installer to silently install the application into the
+    destination folder. Return the folder path.
 
+    Arguments:
+    src -- MSI installer to be executed
+    dest -- the path to install to
+
+    """
+    if os.path.exists(dest):
+        raise InstallError('Installation directory "%s" already exists' % dest)
+
+    # possibly gets around UAC in vista (still need to run as administrator)
     os.environ['__compat_layer'] = 'RunAsInvoker'
     cmd = [src, '/S', '/D=%s' % os.path.realpath(dest)]
 
@@ -228,45 +347,6 @@ def _install_exe(src, dest):
         raise Exception('Execution of installer failed.')
 
     return dest
-
-
-def get_binary(path, apps=_default_apps):
-    """
-    Finds the binary in the specified path
-    path - the path within which to search for the binary
-
-    returns - the full path to the binary in the folder
-              or None if the binary cannot be found
-    """
-
-    binary = None
-
-    # On OS X we can get the real binary from the app bundle
-    if mozinfo.isMac:
-        plist = '%s/Contents/Info.plist' % path
-        assert os.path.isfile(plist), '"%s" has not been found.' % plist
-
-        binary = os.path.join(path, 'Contents/MacOS/',
-                              readPlist(plist)['CFBundleExecutable'])
-
-    else:
-        if mozinfo.isWin:
-            apps = [app + '.exe' for app in apps]
-
-        for root, dirs, files in os.walk(path):
-            for filename in files:
-                # os.access evaluates to False for some reason, so not using it
-                if filename in apps:
-                    binary = os.path.realpath(os.path.join(root, filename))
-                    break
-
-    if not binary:
-        # TODO: It should never happen but if it's the case the installed
-        # files should be removed. It can be done once the uninstaller has
-        # been implemented by bug 757411. For now just throw an exception.
-        raise InvalidBinary('"%s" does not contain a valid binary.' % path)
-
-    return binary
 
 
 def cli(argv=sys.argv[1:]):
@@ -281,7 +361,7 @@ def cli(argv=sys.argv[1:]):
                       help='[optional] Directory to install application into')
     parser.add_option('--app', dest='app',
                       action='append',
-                      default=_default_apps,
+                      default=DEFAULT_APPS,
                       help='[optional] Application being installed. '
                            'Should be lowercase, e.g: '
                            'firefox, fennec, thunderbird, etc.')
